Ajuste nos testes

+++ app/Http/Controllers/TagController.php
@@ -6,7 +6,10 @@ use App\Http\Requests\ListTagsRequest;
 use App\Http\Requests\StoreTagRequest;
 use App\Http\Requests\UpdateTagRequest;
 use App\Http\Resources\TagResource;
+use App\Models\Project;
 use App\Models\Tag;
+use Illuminate\Auth\Events\Validated;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class TagController extends Controller {
 	/**
@@ -19,10 +22,12 @@ class TagController extends Controller {
 	}
 
 	/**
-	 * Store a newly created resource in storage.
+	 * Store a newly created resource in storage
 	 */
 	public function store(StoreTagRequest $request) {
-		$tag = Tag::create($request->all());
+		$project = Project::find($request->project_id);
+
+		$tag = $project->tags()->create($request->all());
 
 		return new TagResource($tag);
 	}
@@ -39,6 +44,9 @@ class TagController extends Controller {
 	 */
 	public function update(UpdateTagRequest $request, Tag $tag) {
 		$tag->update($request->all());
+		if ($tag->project->user_id != request()->user()->id) {
+			throw new NotFoundHttpException('Tag not found');
+		}
 		return new TagResource($tag);
 	}
 
@@ -46,6 +54,9 @@ class TagController extends Controller {
 	 * Remove the specified resource from storage.
 	 */
 	public function destroy(Tag $tag) {
+		if ($tag->project->user_id != request()->user()->id) {
+			throw new NotFoundHttpException('Tag not found');
+		}
 		$tag->delete();
 		return response()->noContent();
 	}
